refactor(update-test-asm): put assembly bytes into separate file
Currently, the bytes for each test are embedded into the .cpp file and
are updated by the update-test-asm script. Experience exposed a few
problems:

* clang-format fights with the formatting of update-test-asm.
* update-test-asm can't easily be included in the build system.
* The embedded bytes get verbose.

Instead of storing the bytes inline in the source code, store them in a
separate file called test-asm-generated.cpp. This new file is understood
by Meson and is updated automatically.